test:  Re-ran explore duplicates notebook.  Concluded that the duplicate issue has been resolved.
Testing in January found systematic duplicates of positions over times.  This was found to be due to a 'feature' of metocean buoys.  They are able to reduce the frequency of their position measurements while continuing to take temperature and pressure data.  When this happens GPSDELAY > 0 minutes indicating stale GNSS data.  These repeat positions were removed from the Database since they are not useful.  If someone is interested in the temperature or pressure data they can find it in the raw data file.